1.移除push_bear 2.登录自动送积分功能改为手动签到送流量 3.将订阅和节点列表独立出来，个人中心首页干净了 4.升级验证码为科学计算
