+++ app/Http/Middleware/VerifyCsrfToken.php
@@ -12,6 +12,7 @@ class VerifyCsrfToken extends Middleware
      * @var array
      */
     protected $except = [
-        "payment/*"
+        "payment/*",
+        "checkIn"
     ];
 }
