+++ resources/views/admin/userList.blade.php
@@ -324,7 +324,7 @@
         // 复制订阅链接
         var clipboard = new Clipboard('.copySubscribeLink');
         clipboard.on('success', function(e) {
-            layer.alert("成功复制该用户的订阅链接", {icon: 1});
+            layer.alert("成功复制该用户的订阅链接", {icon: 1, title:'提示'});
         });
         clipboard.on('error', function(e) {
             console.log(e);
