+++ resources/views/user/buy.blade.php
@@ -75,11 +75,11 @@
             @endif
             <div class="row">
                 <div class="col-xs-12" style="text-align: right;">
-                    @if($is_youzan)
+                    @if(\App\Components\Helpers::systemConfig()['is_youzan'])
                         <a class="btn btn-lg red hidden-print" onclick="onlinePay(0)"> {{trans('home.online_pay')}} </a>
-                    @elseif($is_alipay)
+                    @elseif(\App\Components\Helpers::systemConfig()['is_alipay'])
                         <a class="btn btn-lg green hidden-print" onclick="onlinePay(4)"> 支付宝扫码 </a>
-                    @elseif($is_f2fpay)
+                    @elseif(\App\Components\Helpers::systemConfig()['is_f2fpay'])
                         <a class="btn btn-lg green hidden-print" onclick="onlinePay(5)"> 支付宝扫码 </a>
                     @endif
                   	@if($goods->type <= 2)
