+++ scripts/vnstat.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-yum install -y wget sqlite-devel
+yum install -y wget sqlite-devel gcc
 wget https://github.com/vergoh/vnstat/releases/download/v2.1/vnstat-2.1.tar.gz
 tar zxvf vnstat-2.1.tar.gz
 cd vnstat-2.1
